python/eventhub/requirements.txt
azure-core
azure-eventhub
certifi
chardet
idna
requests
six
uamqp
urllib3
